% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WNN.R
\docType{class}
\name{WNN}
\alias{WNN}
\title{Weighted nearest neighbours for time series forecasting}
\description{
R6 class implementing a weighted nearest neighbours (WNN) algorithm
for univariate time series forecasting. The model uses the last \emph{p}
observations as a window, searches for the \emph{K} most similar historical
windows, and predicts future values as a distance-weighted combination of
the neighbours' subsequent observations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WNN-new}{\code{WNN$new()}}
\item \href{#method-WNN-fit}{\code{WNN$fit()}}
\item \href{#method-WNN-predict}{\code{WNN$predict()}}
\item \href{#method-WNN-clone}{\code{WNN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WNN-new"></a>}}
\if{latex}{\out{\hypertarget{method-WNN-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WNN$new(K, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{The number of neighbours to select}

\item{\code{p}}{The number of past periods to use (window size)
Fit the WNN model to training data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WNN-fit"></a>}}
\if{latex}{\out{\hypertarget{method-WNN-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WNN$fit(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A numeric or ts time series to fit the model
Predict the next values over h periods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WNN-predict"></a>}}
\if{latex}{\out{\hypertarget{method-WNN-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WNN$predict(h)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h}}{The number of periods to predict (forecast horizon)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ts object of length h with the WNN forecast
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WNN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WNN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WNN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
